Make scribes threadsafe.
Added Doodle object that holds mutable state (level/indentation).
Instead of scribe.render() returning a string, it writes into the
doodle so the doodle also holds output. That changed the API
to most callers.

Different threads would presumably be building their own output buffers.
It's up to the code in those threads to decide if/when/how to coordinate
them (just as any logging mechanism would, only here the interleaving can
be at a coarser granularity to meet the needs of the caller).

Whereas before render calls were in the form:
   (obj_to_render, scribe_to_use)
now they are in the form:
   (doodle_to_write_to, obj_to_render)
where the doodle contains the scribe as an attribute making the
calls simpler. "doodle_to_write_to" is typically called "out" for brevity.
Note that the order of parameters changed so that the output parameter is
first. This is for consistency against different APIs that are writing to
the output but take different types of arguments and argument lists. The output
is always positionally first.

Made HtmlScribe's internal unique id counter threadsafe.

Reordered parameters.