Added snapshot and journal packages to start replacing Scribes.
A snapshot captures the state and relationships of the data model at
a particular point in time. This is stored as graph of entities with
annotated edges among them. Where values are trivial, the values are
encoded into the edge itself as a _value annotation.

A journal is just a sequence of snapshots that can write to a file.
In the future, components will be added to render the journal into
a report to facilitate debugging or other reading.

This replaces scribes which tried to encode the report directly, but
that required anticiting up front what should be included in the report
and how it should be rendered, and the components often reference common
objects without being clear that these are the same references. While it
isnt clear yet how to render the JSON, the graphs should provide a more
expressive capture and can allow rendering to be performed later, perhaps
over multiple passes, with different viewpoints to support the task at hand.
And the journal can contain a more complete trace with all the details, letting
the post process reporting pick the level of abstraction and scope to
include in the report.

Whereas in practice scribes used a Scribable interface that took "parts",
and an optional specialized renderer, where each part contained a relation
to the thing it was a part of, the JsonSnapshot uses a "JsonSnapshotable"
interface were the object is given an entity to fill out and the snapshot
that contains the entity. The object can write its components into the snapshot
and build the annotated edges to them, using the snapshot as the data store
so that references to common entities can be shared within the graph. A helper
class within the snapshot for building edges keeps this to what I hope is
a similar ease for snapshoting a typical object.