Use JsonSnapshot throughout the json_contract package and deprecate Scribe.
This is mostly a port of the scribes (_make_scribe_parts method), but I tweaked
the data model a bit, and added more metadata. Also I optimized some
abstractions by flattening out lists that only contained one element to make
the model easier to navigate.

I also removed scribes from the tests, and use the snapshots instead.
This forced me to refactor how I was constructing some expected values.
The tests and values are the same, but the actual data model is different.
The snapshots capture actual references, not value references. So equivalent
values are not enough for the test to pass. The reasons the values are
equivalent (internally referencing the same object) is also implied by the
snapshot.

Also some whitespace formatting changes to please pylint.