Make JsonSnapshotable a bit more generic.
Renamed JsonSnapshotable to JsonSnapshotableEntity meaning that the JSON
representation is an entity (composite value). Added JsonSnapshotable base
class with a method that returns a JSON value that can be added into a snapshot
representation.

The difference is that objects can be made "JsonSnapshotable" that return
simple values rather than composite values. This will allow wrapper objects
(for edge values) that can return simple values without having to expose the
wrapper to the snapshot.

This is also a bit simpler API, even just for JsonSnapshotableEntity.

Also, renamed the snapshot _for_data methods as _for_object because
they refer to "entities" (which are composite objects) and not simple
data objects (e.g. wrapped values). Simple objects arent added to snapshots,
though are used for edge attributes describing relationships between objects
in a snapshot.