Changed operation status_collector and cleanup functions.
The old status_collector (and cleanup) functions predated contexts.
Changed status_collector to status_extractor, whose responsibility can
be to move values from the status into the context. in the case of spinnaker,
this could traverse the complex json data to extract out a couple individual
fields that might be referenced by a test.

Cleanup just takes a context. The run_test_case method adds the verify results
into the context so they are still available (and the status was already there).

I also renamed the keys that the run_test_case writes into the context to be
camel case to match the type names. I dont think anything was referencing this
yet.