Breaking change to spinnaker.binding key names.
This is a first PR preparing to migrate out the spinnaker tests. I
updated some docs and renamed some of the parameters and bindings to
be less daunting and clearer what the parameters are for.

The changes to binding dictionary keys is a breaking change. I'm not
aware of any significant inconvienience this will cause people
(e.g. PRs in progress) The breaking changes are straight forward:
<PLATFORM>_CREDENTIALS -> SPINNAKER_<PLATFORM>_ACCOUNT

The change in parameters is a deprecated change, maintaining support
for the old interface. I'll keep these around of a while since it is
easy to support.