Added rendering transforms to simplify report structure
The data model dumps show the literal data model. There are some
common types that use lists for a general case so these are reported
out as lists even if the particular instance is only dealing with one
element.

If the list is just an implementation detail (as opposed to something
that is fundamentally a list and may contain one) then this list rendering
adds abstraction overhead and complexity for the reader. This PR allows
certain types to explicity transform the interpretation of their encodings
depending on the context of the instance value.

For example PathValueResult contains a list of values at a given path
because in general paths can pass through lists as intermediate nodes,
so have different branches down the path resulting in different values.
If there is no list, then the PathValue is a list of 1, but that is just
an implementation detail. Conceptually  the path value is the element in
that containing list so should be reported that way.

The transforms are on the reporting side rather than on the snapshot side.
This might be a mistake, but I think that there could be other reasons to
read snapshots in the future so having a consistent data model will be
simpler. The transforms on the interpretation are not tied to the HTML
rendering, so all renderers (and readers of journal data for that matter)
could share the same implementation as if it were in the snapshot anyway.