Journal the whole test fixture method, not just the OperationContract executions
When only journaling operation contract, lower level errors are left out
(e.g. exceptions in setUp and tearDown). This also fixes logging/journaling so
that logs are visible earlier than before.

The downside of this is that it adds another level of abstraction -- the
test method is often just a wrapper around the OperationContract. However
breaking up the test logging flow to distinguish setUp, Execution, and
tearDown seems useful, particularly when there are errors or lazy initialization
in setUp, which now gets scoped to the test it is for as well as separated from
the execution of the text.

To clean up reporting, the HTML reporter is specialized to ommit empty
setUp and tearDown context blocks and to promote the Execution context to
the parent (test). The net effect is that a test will look like this:
     TestFunctionName
         setUp
            trace from fixture.setUp
         trace from actual test
         tearDown
            trace from fixture.tearDown

or maybe like this if there is no setUp or tearDown
     TestFunctionName
         trace from actual test

I'd like to get rid of the outer TestFunctionName where there is only an
OperationContract but that would require a larger refactoring.