Added ConfigurationBindings management. (#237)
Use ConfigurationBindings for the bindings instead of the argparser.
Deprecated initArgumentParser in favor of init_bindings_builder.
The ConfigurationBindingsBuilder can also create command-line arguments
so that bindings can be replaced either by command-line or config files.

With the deprecation, this is fully backward compatible with the existing
tests. I'm going to migrate the tests over in another PR and give other
authors a chance to migrate before I take out the initArgumentParser api.