fix(AgentTestScenario): No longer allow initParser with ArgumentParser (#251)
This is a breaking change except that the call was deprecated a long time
ago and is not used with ArgumentParser anymore. The method itself is still
needed because most of the Spinnaker tests use it for initialization. However
it is called with a ConfigurationBindingsBuilder as part of the deprecation.

The real point of this PR is to fix the initialization flow to ensure that
init_bindings_builder is always called to prepare for some additional
refactoring of how bindings are initialized.