feat(validate): Adds ObservationPredicateFactory and GCP refactoring. (#271)
BREAKING CHANGE

This converts over the gcp_testing to the new API,
especially handling observation errors. These are no longer
options into the observer method, rather they are specified
by the caller's ObservationVerifier (such as the
new ObservationErrorVerifier).

I converted over the spinnaker tests in another PR. I'm not aware
of other use cases that tested for errors the way removed so this
so am removing the obsolete implementation here and now as well
making this a breaking change.

This PR also adds a factory class for creating predicates to complement
ObservationVerifier EXPECT/OR/AND methods so as to keep that
api more lean. This is still more unweildy than I'd like, but
is more consistent and clearer than the original ValueObservationVerifier
API was.

I could add more specialized constructor varianets but am holding off
on that. At least now it is more clear and consistent to use.