fix(reporting): Fixed reporting to convey fixture errors outside tests
Tests can fail if there is an exception outside a test, however the
citest journals only recorded success and failures of the tests themselves.
This resulted in tests failing when there is a classTearDown error, but
the test reports (and index) not indicating this.

The fix for this is in several places:
  (1) A snapshot of the fixture unittest.TestResult is written at the end
      of the journal. The unittest.TestResult already captures these errors
      (however the errors are caught inside the base unittest classes
       outside the view of the citest runner or specializations)

  (2) The index generator looks for this and uses the relation hint on
      the snapshot to influence how it marks the overall test status

  (3) The HTML renderer looks for a relation hint to color snapshot
      summaries so that these errors will become visible in the normal
      reports.