chore(reporting): Move JournalNavigator and JournalProcessor to base
These used to be in the reporting package, but have no dependencies
other than journal. Moving them into the base package to be near
journal, especially for test support.

I'd move these into a "journal" package along with journal and
snapshoting, none of which depend on anything else in base, however
journal is supposed to be ubiquotous so seems reasonable that future
features and refactorings might introduce dependencies.

The reporting package is still where interpretation and rendering is.